Update index.js to handle local files
I was able to get the PDFWindow to work perfectly for online (http/https) files, but when attempting to open a local file I got a prompt to save instead of opening. After looking through the code, i noticed there was a "/" missing from the local file:/// portion of the Promise.